+++ modules/__tests__/HashHistory-test.js
@@ -69,7 +69,7 @@ describeHistory('a hash history', () => {
     })
 
     describe('push with an encoded path string', () => {
-      it('creates a location object with decoded pathname', (done) => {
+      it('creates a location object with decoded pathname and rawPathname', (done) => {
         TestSequences.PushEncodedLocation(history, done)
       })
     })